renamed modules to remove plurals
