+++ test/unittests/interface/test_mrtrix.py
@@ -1,7 +1,7 @@
 import os.path
 from nianalysis.testing import BaseTestCase
 import tempfile
-from nianalysis.nodes import Node
+from nianalysis.node import Node
 from nianalysis.interfaces.mrtrix import MRCalc
 from mbianalysis.requirements import mrtrix3_req
